Move new member form into its own component
